fixed preauthorization bug

+++ src/main/java/hu/bme/edu/handmade/controllers/UserController.java
@@ -1,5 +1,5 @@
 package hu.bme.edu.handmade.controllers;
-
+import hu.bme.edu.handmade.exception.ForbiddenException;
 import hu.bme.edu.handmade.mappers.AddressMapper;
 import hu.bme.edu.handmade.mappers.UserMapper;
 import hu.bme.edu.handmade.models.CartProduct;
@@ -26,6 +26,7 @@ import java.net.URLDecoder;
 import java.nio.charset.StandardCharsets;
 import java.security.Principal;
 import java.util.List;
+import java.util.Objects;
 import java.util.Optional;
 import java.util.UUID;
 
@@ -87,10 +88,14 @@ public class UserController {
         }
     }
 
-    @PreAuthorize("hasAuthority('ROLE_USER') && #id == principal.getUsername()")
+    @PreAuthorize("hasAuthority('ROLE_USER')")
     @SecurityRequirement(name = "Bearer_Authentication")
     @PutMapping("/{id}")
-    public ResponseEntity<User> updateUser(@PathVariable("id") Long id, @RequestBody UserDto userToUpdate) {
+    public ResponseEntity<User> updateUser(@PathVariable("id") Long id, @RequestBody UserDto userToUpdate, Principal principal) {
+        User principalUser = userService.findUserByEmail(principal.getName());
+        if (!Objects.equals(principalUser.getId(), id)) {
+            throw new ForbiddenException();
+        }
         Optional<User> updatedUser = userService.updateUser(id, userToUpdate);
 
         return updatedUser
@@ -104,61 +109,92 @@ public class UserController {
                     return ResponseEntity.created(location).body(createdUser);
                 });
     }
-    @PreAuthorize("hasAuthority('ROLE_USER') && #userId == principal.getUsername()")
+    @PreAuthorize("hasAuthority('ROLE_USER')")
     @SecurityRequirement(name = "Bearer_Authentication")
     @PostMapping("/{id}/addresses")
-    public AddressDto addUserAddress(@PathVariable("id") Long userId, @RequestBody AddressDto address) {
+    public AddressDto addUserAddress(@PathVariable("id") Long userId, @RequestBody AddressDto address, Principal principal) {
+        User principalUser = userService.findUserByEmail(principal.getName());
+        if (!Objects.equals(principalUser.getId(), userId)) {
+            throw new ForbiddenException();
+        }
         return AddressMapper.INSTANCE.toAddressDto(userService.addAddress(userId, address));
     }
 
-    @PreAuthorize("hasAuthority('ROLE_USER') && #userId == principal.getUsername()")
+    @PreAuthorize("hasAuthority('ROLE_USER')")
     @SecurityRequirement(name = "Bearer_Authentication")
     @PutMapping("/{id}/addresses/{addressId}")
     public AddressDto updateUserAddress(@PathVariable("id") Long userId,
                                         @PathVariable("addressId") Long addressId,
-                                        @RequestBody AddressDto address) {
+                                        @RequestBody AddressDto address,
+                                        Principal principal) {
+        User principalUser = userService.findUserByEmail(principal.getName());
+        if (!Objects.equals(principalUser.getId(), userId)) {
+            throw new ForbiddenException();
+        }
         return AddressMapper.INSTANCE.toAddressDto(userService.updateAddress(userId, address));
     }
 
     /** Newsletter subscription */
-    @PreAuthorize("hasAuthority('ROLE_USER') && #userId == principal.getUsername()")
+    @PreAuthorize("hasAuthority('ROLE_USER')")
     @SecurityRequirement(name = "Bearer_Authentication")
     @PostMapping("/{id}/unsubscribe")
-    public void unsubscribeFromNewsletter(@PathVariable("id") Long userId) {
+    public void unsubscribeFromNewsletter(@PathVariable("id") Long userId, Principal principal) {
+        User principalUser = userService.findUserByEmail(principal.getName());
+        if (!Objects.equals(principalUser.getId(), userId)) {
+            throw new ForbiddenException();
+        }
         userService.setNewsletterSubscription(userId, false);
     }
 
     /** Cart */
-    @PreAuthorize("hasAuthority('ROLE_USER') && #userId == principal.getUsername()")
+    @PreAuthorize("hasAuthority('ROLE_USER')")
     @SecurityRequirement(name = "Bearer_Authentication")
     @GetMapping("/{id}/cart")
-    List<CartProduct> getCartProducts(@PathVariable("id") long userId) {
+    List<CartProduct> getCartProducts(@PathVariable("id") long userId, Principal principal) {
+        User principalUser = userService.findUserByEmail(principal.getName());
+        if (!Objects.equals(principalUser.getId(), userId)) {
+            throw new ForbiddenException();
+        }
         return cartService.getCartProductsByUser(userId);
     }
 
-    @PreAuthorize("hasAuthority('ROLE_USER') && #userId == principal.getUsername()")
+    @PreAuthorize("hasAuthority('ROLE_USER')")
     @SecurityRequirement(name = "Bearer_Authentication")
     @PostMapping("/{id}/cart")
-    CartProduct addCartProduct(@PathVariable("id") long userId, @RequestBody CartProductDto cartProductDto) {
+    CartProduct addCartProduct(@PathVariable("id") long userId, @RequestBody CartProductDto cartProductDto, Principal principal) {
+        User principalUser = userService.findUserByEmail(principal.getName());
+        if (!Objects.equals(principalUser.getId(), userId)) {
+            throw new ForbiddenException();
+        }
         return cartService.addCartProduct(cartProductDto, userId);
     }
 
-    @PreAuthorize("hasAuthority('ROLE_USER') && #userId == principal.getUsername()")
+    @PreAuthorize("hasAuthority('ROLE_USER')")
     @SecurityRequirement(name = "Bearer_Authentication")
     @PutMapping("/{userId}/cart/{cartId}")
     CartProduct updateCartProduct(@PathVariable("userId") long userId,
                                   @PathVariable("cartId") long cartId,
-                                  @RequestBody CartProductDto cartProductDto) {
+                                  @RequestBody CartProductDto cartProductDto,
+                                  Principal principal) {
+        User principalUser = userService.findUserByEmail(principal.getName());
+        if (!Objects.equals(principalUser.getId(), userId)) {
+            throw new ForbiddenException();
+        }
         return cartService.findCartProductById(cartId)
                 .map(product -> cartService.updateCartProduct(cartProductDto, cartId))
                 .orElseGet(()->cartService.addCartProduct(cartProductDto, userId));
     }
 
-    @PreAuthorize("hasAuthority('ROLE_USER') && #userId == principal.getUsername()")
+    @PreAuthorize("hasAuthority('ROLE_USER')")
     @SecurityRequirement(name = "Bearer_Authentication")
     @DeleteMapping("/{userId}/cart/{cartId}")
     ResponseEntity<?> removeCartProduct(@PathVariable("userId") Long userId,
-                                        @PathVariable("cartId") Long cartId) {
+                                        @PathVariable("cartId") Long cartId,
+                                        Principal principal) {
+        User principalUser = userService.findUserByEmail(principal.getName());
+        if (!Objects.equals(principalUser.getId(), userId)) {
+            throw new ForbiddenException();
+        }
         try {
             cartService.deleteCartProduct(cartId);
             return ResponseEntity.noContent().build();
@@ -169,26 +205,40 @@ public class UserController {
     }
 
     /** Wishlist */
-    @PreAuthorize("hasAuthority('ROLE_USER') && #userId == principal.getUsername()")
+    @PreAuthorize("hasAuthority('ROLE_USER')")
     @SecurityRequirement(name = "Bearer_Authentication")
     @GetMapping("/{user_id}/wishlist")
-    List<ProductDto> getWishList(@PathVariable("user_id") Long id) {
+    List<ProductDto> getWishList(@PathVariable("user_id") Long id, Principal principal) {
+        User principalUser = userService.findUserByEmail(principal.getName());
+        if (!Objects.equals(principalUser.getId(), id)) {
+            throw new ForbiddenException();
+        }
         return wishListService.getWishListItemsByUserId(id);
     }
 
-    @PreAuthorize("hasAuthority('ROLE_USER') && #userId == principal.getUsername()")
+    @PreAuthorize("hasAuthority('ROLE_USER')")
     @SecurityRequirement(name = "Bearer_Authentication")
     @PostMapping("/{user_id}/wishlist")
     ProductDto addToWishList(@PathVariable("user_id") Long userId,
-                             @RequestParam("product_id") Long productId) {
+                             @RequestParam("product_id") Long productId,
+                             Principal principal) {
+        User principalUser = userService.findUserByEmail(principal.getName());
+        if (!Objects.equals(principalUser.getId(), userId)) {
+            throw new ForbiddenException();
+        }
         return wishListService.addToWishlist(userId, productId);
     }
 
-    @PreAuthorize("hasAuthority('ROLE_USER') && #userId == principal.getUsername()")
+    @PreAuthorize("hasAuthority('ROLE_USER')")
     @SecurityRequirement(name = "Bearer_Authentication")
     @DeleteMapping("/{user_id}/wishlist")
     void removeFromWishList(@PathVariable("user_id") Long userId,
-                            @RequestParam("product_id") Long productId) {
+                            @RequestParam("product_id") Long productId,
+                            Principal principal) {
+        User principalUser = userService.findUserByEmail(principal.getName());
+        if (!Objects.equals(principalUser.getId(), userId)) {
+            throw new ForbiddenException();
+        }
         wishListService.removeFromWishList(userId, productId);
     }
